requirements.txt
fastapi==0.109.0
pydantic==2.12.5
python-dotenv==1.2.1
structlog==25.5.0
rich==14.2.0
uvicorn==0.38.0
SQLAlchemy==2.0.45
asyncpg==0.31.0
